% modification on git from copied files
\name{hSL}
\docType{data}
\alias{hSL}

\title{Preprocessed 'lociData' object containing likelihoods of methylation
at each locus.}

\description{This is the preprocessed (to save computational time)
  'lociData' object described in the vignette 'segmentSeq: methods for
  detecting methylation loci and differential methylation'.}

\usage{hSL}
\format{A 'lociData' object.}
\source{See vignette.}

\keyword{datasets}
    