\name{SL}
\docType{data}
\alias{SL9}
\alias{SL10}
\alias{SL26}
\alias{SL32}

\title{Example data selected from a set of Illumina sequencing
  experiments.}

\description{Each of the files 'SL9', 'SL10', 'SL26' and 'SL32'
  represents a subset of the data from an Illumina sequencing
  experiment. These data consist of alignment information; the tag
  sequence, and the number of times that each sequence is observed.
}

\usage{SL}
\format{A set of tab-delimited files containing data from four
  sequencing experiments.}
\source{In-house Illumina sequencing experiments}


\keyword{datasets}
    